pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "llamatrader-openenv"
version = "0.1.0"
description = "An OpenEnv compliant algorithmic daytrading simulator"
readme = "README.md"
authors = [
    {name = "Siddha Narayan Patra"}
]
requires-python = ">=3.12"

# You can leave this empty if you are relying on requirements.txt for Docker, 
# but OpenEnv looks for this section to exist.
dependencies = [
    "openenv-core"
]

[project.scripts]
server = "server.app:main"